[Android] Move parentId from CriticalPersistedTabData back to Tab
Chromium change:
https://github.com/brave/chromium/commit/03e01235adfb1949dba563390ca9bf1450b0dbdf

Move parentId from CriticalPersistedTabData back to Tab
parentId is moved back to Tab from CriticalPersistedTabData,
reversing the usage of the UserData pattern. This is part of
CriticalPersistedTabData deprecation.
CriticalPersistedTabData#getParentId()
is replaced with a placeholder value in the persistence parts
of CriticalPersistedTabData. This is safe to do since the experiment
is off. This code will also be removed as part of
CriticalPersistedTabData deprecation. It hasn't been done in
this CL to keep CLs focused and tractable.

Bug: b/300339871